Begin creating a base list component class.
