+++ src/components/gear-list/gear-list.html
@@ -17,7 +17,7 @@
         rotate: 180deg;
     }
 </style>
-<table id="gear-list-container" class="w-full mb-18 text-left border-collapse border-3 border-black">
+<table id="list-container" class="w-full mb-18 text-left border-collapse border-3 border-black">
     <thead class="text-white bg-black"></thead>
     <tbody></tbody>
 </table>
