+++ src/components/npc-list/npc-list.html
@@ -1,4 +1,4 @@
-<table id="npc-list-container" class="w-full mb-18 text-left border-collapse border-3 border-black">
+<table id="list-container" class="w-full mb-18 text-left border-collapse border-3 border-black">
     <thead class="text-white bg-black"></thead>
     <tbody></tbody>
 </table>
