+++ src/components/npc-list/npc-list.ts
@@ -1,5 +1,4 @@
 import html from "./npc-list.html";
-import { BaseComponent } from "../base.component";
 import { IUnitOfWork } from "../../lib/common/data-access/unit-of-work-interface";
 import { appInjector } from "../../lib/infrastructure/app-injector";
 import { UnitOfWork } from "../../lib/data-access/unit-of-work";
@@ -24,23 +23,20 @@ import { NpcFilterChangedEvent } from "../../lib/events/npc-filter-changed-event
 import { AppEventListener } from "../../lib/events/app-event-listener-interface";
 import { FilterNpcsListFeature } from "../../features/npcs/filter-npcs-list/filter-npcs-list-feature";
 import { FilterNpcsListRequest } from "../../features/npcs/filter-npcs-list/filter-npcs-list-request";
+import { BaseListComponent } from "../base-list/base-list-component";
 
-export class NpcListComponent extends BaseComponent {
+export class NpcListComponent extends BaseListComponent {
     private getAllNpcsFeature: GetAllNpcsFeature;
     private npcsList: Array<NpcListItem>;
-    private tableHeaders: TableHeader[] = [
-        new TableHeader(SortNpcsListFeature.fieldName, "Name"),
-        new TableHeader(SortNpcsListFeature.fieldCombat, "Combat"),
-        new TableHeader(SortNpcsListFeature.fieldInstinct, "Instinct"),
-        new TableHeader(SortNpcsListFeature.fieldArmorPoints, "Armor Points"),
-        new TableHeader(SortNpcsListFeature.fieldWoundsHealth, "Max Wounds (Health)"),
-    ];
-    private sortState: SortState = new SortState(SortNpcsListFeature.fieldId);
-    private unitOfWork: IUnitOfWork;
 
     constructor() {
-        super();
-        this.unitOfWork = appInjector.injectClass(UnitOfWork);
+        super(SortNpcsListFeature.fieldId, [
+            new TableHeader(SortNpcsListFeature.fieldName, "Name"),
+            new TableHeader(SortNpcsListFeature.fieldCombat, "Combat"),
+            new TableHeader(SortNpcsListFeature.fieldInstinct, "Instinct"),
+            new TableHeader(SortNpcsListFeature.fieldArmorPoints, "Armor Points"),
+            new TableHeader(SortNpcsListFeature.fieldWoundsHealth, "Max Wounds (Health)"),
+        ]);
         this.getAllNpcsFeature = new GetAllNpcsFeature(this.unitOfWork);
         this.npcsList = [];
     }
@@ -50,7 +46,7 @@ export class NpcListComponent extends BaseComponent {
 
         this.npcsList = this.getAllNpcsFeature.handle(new EmptyRequest());
 
-        this.sortNpcs();
+        this.sortItems();
 
         this.populateTableHeaderRow();
         this.populateTableRows();
@@ -62,53 +58,8 @@ export class NpcListComponent extends BaseComponent {
         EventBus.instance.unregister(NpcFilterChangedEvent.name, this.onNpcCategoryChangedEvent);
     }
 
-    private populateTableHeaderRow() {
-        const npcListContainer = this.shadow.querySelector("#npc-list-container");
-
-        if (!npcListContainer) return;
-
-        const tableHeader = npcListContainer.querySelector("thead");
-
-        if (!tableHeader) return;
-
-        tableHeader.appendChild(this.createTableHeaderRowElement());
-    }
-
-    private createTableHeaderRowElement(): HTMLTableRowElement {
-        const row = document.createElement("tr");
-
-        this.tableHeaders.forEach((header) => {
-            row.appendChild(this.createTableHeaderElement(header));
-        });
-
-        return row;
-    }
-
-    private createTableHeaderElement(header: TableHeader): HTMLTableCellElement {
-        const tableHeaderElement = document.createElement("th");
-
-        tableHeaderElement.id = `header${header.field}`;
-        tableHeaderElement.className = "uppercase p-2 cursor-pointer";
-        tableHeaderElement.onclick = () => this.onTableHeaderClick(header.field);
-        tableHeaderElement.appendChild(this.createHeaderDivElement(header));
-
-        return tableHeaderElement;
-    }
-
-    private createHeaderDivElement(header: TableHeader): HTMLDivElement {
-        const headerDiv = document.createElement("div");
-
-        headerDiv.className = "flex justify-between";
-        headerDiv.innerHTML = `
-            <div>${header.displayName}</div>
-            <div id="${header.field}SortIcon"></div>
-        `;
-
-        return headerDiv;
-    }
-
-    private populateTableRows(clearOldRows: boolean = false) {
-        const npcListContainer = this.shadow.querySelector("#npc-list-container");
+    protected populateTableRows(clearOldRows: boolean = false) {
+        const npcListContainer = this.shadow.querySelector("#list-container");
 
         if (!npcListContainer) return;
 
@@ -153,13 +104,7 @@ export class NpcListComponent extends BaseComponent {
         return value.toString();
     }
 
-    public onTableHeaderClick(header: string) {
-        this.sortState.set(header);
-        this.sortNpcs();
-        this.populateTableRows(true);
-    }
-
-    private sortNpcs() {
+    protected sortItems() {
         this.tableHeaders.forEach((currentHeader) => {
             this.updateSortIcons(currentHeader.field);
         });
@@ -178,25 +123,6 @@ export class NpcListComponent extends BaseComponent {
         }
 
         this.npcsList = result.value ?? [];
-    }
-
-    private updateSortIcons(currentHeader: string) {
-        const sortIcon = this.shadow.querySelector(`#${currentHeader}SortIcon`);
-
-        if (!sortIcon) return;
-
-        if (currentHeader !== this.sortState.field) {
-            sortIcon.className = "";
-            return;
-        }
-
-        if (this.sortState.direction === SortDirection.Ascending) {
-            sortIcon.className = "icon-chevron icon-chevron-up";
-        } else if (this.sortState.direction === SortDirection.Descending) {
-            sortIcon.className = "icon-chevron icon-chevron-down";
-        } else {
-            sortIcon.className = "";
-        }
     }
 
     public onTableDataRowClick(npcListItem: NpcListItem) {
@@ -247,13 +173,9 @@ export class NpcListComponent extends BaseComponent {
 
         this.npcsList = result.value ?? [];
 
-        this.sortNpcs();
+        this.sortItems();
 
         this.populateTableRows(true);
-    }
-
-    private dispatchErrorEvent(error: string) {
-        EventBus.instance.dispatch(new AppErrorEvent(EventType.ErrorPanelShow, error));
     }
 }
